:soap: lint tests
